+++ packages/helpers/src/index.ts
@@ -109,7 +109,12 @@ export {
   formatDuration,
   parseDate,
 } from "./date";
-export { humanizeSeconds } from "./duration";
+export {
+  humanizeSeconds,
+  DURATION_UNITS,
+  toMaxAgeSeconds,
+  fromMaxAgeSeconds,
+} from "./duration";
 export { getTrustCenterUrl } from "./trustCenter";
 export { detectSocialName } from "./socialUrl";
 export { formatError, type GraphQLError } from "./error";
